+++ src/Processors/QueryPlan/ReadFromSystemNumbersStep.cpp
@@ -41,7 +41,7 @@ protected:
         auto column = ColumnUInt64::create(block_size);
         ColumnUInt64::Container & vec = column->getData();
 
-        size_t curr = next; /// The local variable for some reason works faster (>20%) than member of class.
+        UInt64 curr = next; /// The local variable for some reason works faster (>20%) than member of class.
         UInt64 * pos = vec.data(); /// This also accelerates the code.
         UInt64 * end = &vec[block_size];
         iota(pos, static_cast<size_t>(end - pos), curr);
